small fixes for latest file structure
