+++ bentoctl_sagemaker/delete.py
@@ -1,4 +1,4 @@
-from utils import console, run_shell_command
+from .utils import console, run_shell_command
 from .generate_resource_names import generate_resource_names
 
 def delete(deployment_name, deployment_spec):
